Spliting functions into their own separate files
